packages/mockcreator/requirements.txt
pandas
faker
pyyaml